Handle split datasets in dummy datasets
