+++ cdisc_rules_engine/services/data_services/dummy_data_service.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 from io import IOBase
-from typing import List, Optional, Callable
+from typing import List, Optional, Callable, Iterator
 
 import pandas as pd
 
@@ -124,7 +124,17 @@ class DummyDataService(BaseDataService):
         return True
 
     def join_split_datasets(self, func_to_call: Callable, dataset_names, **kwargs):
-        pass
+        drop_duplicates: bool = kwargs.pop("drop_duplicates", False)
+        datasets: Iterator[pd.DataFrame] = self._async_get_datasets(
+            func_to_call, dataset_names, **kwargs
+        )
+        joined_dataset: pd.DataFrame = pd.concat(
+            [dataset for dataset in datasets],
+            ignore_index=True,
+        )
+        if drop_duplicates:
+            joined_dataset.drop_duplicates()
+        return joined_dataset
 
     def read_data(self, file_path: str) -> IOBase:
         pass
